+++ demo/src/demo.ts
@@ -3,13 +3,15 @@ import { CanvasInfo, initializeCanvas } from 'Canvas/index'
 import { animationDemo } from './animationDemo'
 import { ellipseDemo } from './ellipseDemo'
 import { lineDemo } from './lineDemo'
+import { scalingDemo } from './scalingDemo'
 
 enum Scene {
   StaticDemo,
   AnimationDemo,
+  ScaleDemo,
 }
 
-let scene: Scene = Scene.StaticDemo
+let scene: Scene = Scene.StaticDemo // default scene
 
 function drawScene(ci: CanvasInfo): void {
   switch (scene) {
@@ -20,6 +22,9 @@ function drawScene(ci: CanvasInfo): void {
     case Scene.AnimationDemo:
       animationDemo(ci)
       break
+    case Scene.ScaleDemo:
+      scalingDemo(ci)
+      break
     default:
       console.error('Unknown demo')
   }
@@ -29,19 +34,27 @@ function demo() {
   const ci = initializeCanvas('canvas', drawScene)
   ci.showDebugBox()
 
-  document.getElementById('static-demo')?.addEventListener('click', () => {
-    if (scene !== Scene.StaticDemo) {
-      scene = Scene.StaticDemo
-      ci.stopLoop()
-      ci.redraw()
-    }
+  function onSceneElementClick(elemendId: string, newScene: Scene, callback: () => void) {
+    document.getElementById(elemendId)?.addEventListener('click', () => {
+      if (scene !== newScene) {
+        scene = newScene
+        callback()
+      }
+    })
+  }
+
+  onSceneElementClick('static-demo', Scene.StaticDemo, () => {
+    ci.stopLoop()
+    ci.redraw()
   })
 
-  document.getElementById('animation-demo')?.addEventListener('click', () => {
-    if (scene !== Scene.AnimationDemo) {
-      scene = Scene.AnimationDemo
-      ci.startLoop()
-    }
+  onSceneElementClick('animation-demo', Scene.AnimationDemo, () => {
+    ci.startLoop()
+  })
+
+  onSceneElementClick('scale-demo', Scene.ScaleDemo, () => {
+    ci.stopLoop()
+    ci.redraw()
   })
 }